initial commit of SPARQL endpoint
An initial commit of files to faciliate a SPARQL endpoint for PheKnowLator data. The endpoint uses the DBCLS sparql-proxy web app as the front end and serves triples from a Blazegraph backend.

Note: this currently host cl.owl and not the pkt data. Once the bucket path for the pkt data is known, it needs to be added to the Dockerfile.